bug fixing and closure implementation

+++ src/core/math.odin
@@ -1,7 +1,7 @@
 package aincore
 
-gcd :: proc(a: i32, b: i32) -> i32 {
-    result: i32 = min(a, b)
+gcd :: proc(a: i64, b: i64) -> i64 {
+    result: i64 = min(a, b)
     for result > 0 {
         if a % result == 0 && b % result == 0 {
             break
